sort 5 ok

+++ srcs/smart_moves.c
@@ -6,7 +6,7 @@
 /*   By: framos-p <framos-p@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 16:32:36 by framos-p          #+#    #+#             */
-/*   Updated: 2022/11/19 13:05:22 by framos-p         ###   ########.fr       */
+/*   Updated: 2022/11/20 17:13:55 by framos-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 size_t	find_smallest(t_stack **a)
 {
 	t_stack	*temp;
-	int		i;
+	size_t		i;
 
 	temp = *a;
 	i = 0;
@@ -32,7 +32,7 @@ size_t	find_smallest(t_stack **a)
 size_t	find_biggest(t_stack **a)
 {
 	t_stack	*temp;
-	int		i;
+	size_t		i;
 
 	temp = *a;
 	i = 0;
